Refactor summary to have summary module with dataframe and plot submodules.
