+++ thisnotthat/__init__.py
@@ -1,4 +1,3 @@
-import itertools as it
 from .bokeh_plot import BokehPlotPane
 from .label_editor import LegendWidget, LabelEditorWidget
 from .data_viewer import DataPane, SimpleDataPane
@@ -12,8 +11,8 @@ from .instance_viewer import InformationPane
 from .deckgl_plot import DeckglPlotPane
 from .plot_controls import PlotControlWidget
 from .map_cluster_labelling import JointVectorLabelLayers, MetadataLabelLayers
-from .summarizer_plot import PlotSummaryPane
-from .summarizer_dataframe import DataSummaryPane
+from thisnotthat.summary.plot import PlotSummaryPane
+from thisnotthat.summary.dataframe import DataSummaryPane
 from .map_cluster_labelling import (
     JointVectorLabelLayers,
     MetadataLabelLayers,
